Send Data

+++ give_conn.py
@@ -16,11 +16,11 @@ def delete_connections(r):
 
 try:
     while True:
-        if r.get('connection_delete_requested').decode() == 'True':
+        if r.get('connection_delete_requested') is not None and r.get('connection_delete_requested').decode() == 'True':
             print("Handling Delete")
             delete_connections(r)
 
-        if r.get('connection_requested').decode() == 'True':
+        if r.get('connection_requested') is not None and r.get('connection_requested').decode() == 'True':
             delete_connections(r)
             print("Handling Create")
             r.set('ssh_conn_url', ngrok.connect(22, "tcp"))
